pyproject.toml
[project]
name = "melby-sama"
version = "0.1.0"
description = ""
authors = [
    {name = "Sodiumchloridy",email = "67376832+Sodiumchloridy@users.noreply.github.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "google-genai (>=0.5.0,<0.6.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "azure-cognitiveservices-speech (>=1.42.0,<2.0.0)",
    "mss (>=10.0.0,<11.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pytchat (>=0.5.5,<0.6.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false